made tests common
